transcribe/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch>=2.1.1
python-multipart==0.0.6
aio_pika